Bootstrapping AppModule.
Typically you bootstrap an Angular app by importing your root module into a new file, main.ts. As Angular can be bootstrapped in multiple environments, such as server-side, we also need to import a module specific to the environment we’re bootstrapping in. For browser bootstrapping, we need to import something called the platform-browser-dynamic module: